+++ src/services/escolas/Associacao.service.js
@@ -45,10 +45,6 @@ export const consultarCodEol = async (cod_eol) => {
     return (await api.get(`/api/membros-associacao/codigo-identificacao/?codigo-eol=${cod_eol}&associacao_uuid=${localStorage.getItem(ASSOCIACAO_UUID)}`, authHeader))
 };
 
-export const consultarNomeResponsavel = async (nome) => {
-    return (await api.get(`/api/membros-associacao/nome-responsavel/?nome=${nome}&associacao_uuid=${localStorage.getItem(ASSOCIACAO_UUID)}`, authHeader))
-};
-
 export const consultarCpfResponsavel = async (cpf) => {
     return (await api.get(`/api/membros-associacao/cpf-responsavel/?cpf=${cpf}&associacao_uuid=${localStorage.getItem(ASSOCIACAO_UUID)}`, authHeader))
 };
